Updates Test of Logout Feature (works now)
This (in addition to the previous pull request) allows the test of user log out functionality to work as expected.  Before these additions, the code passes the user log out test when it's not actually working.  The /sessions/logout route redirects to the home page and the home page has the content "Log In" on it, even though the session hash hasn't been cleared.  This change goes along with a change to home.erb that adds a paragraph at the end of the page containing "Nobody is currently logged in." if session[:id].nil?